feat: rename + extend MAV deploy script to wire UPV-V2 and hand off ownership
- Renames deploy script 23_ → 24_ (23_ was a numbering conflict with
  23_redeploy_whitelist_pre_intent_hook.ts which landed on main via PR #152).
- After deploying MultiAttestationVerifier, the script now also calls
  UnifiedPaymentVerifierV2.setAttestationVerifier(...) and transfers MAV
  ownership to the network's multisig.
- On networks where the deployer still owns UPV-V2 (localhost, base_staging)
  the wiring executes as a direct transaction. On base (prod) UPV-V2 is
  owned by the multisig so the call is appended to safeBatchCollector and
  deploy_summary.ts writes a Safe Transaction Builder batch for later
  multisig execution.
